task1/src/main/java/org/example/service/BookService.java
package org.example.service;

import lombok.RequiredArgsConstructor;
import org.example.model.Book;
import org.example.repository.BookRepository;
import org.springframework.stereotype.Service;
import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class BookService {
	private final BookRepository bookRepository;

	public List<Book> getAllBooks() {
		List<Book> books = new ArrayList<>();
		bookRepository.findAll().forEach(books::add);
		return books;
	}

	public Optional<Book> getBookById(Long id) {
		return bookRepository.findById(id);
	}

	public Book createBook(Book book) {
		return bookRepository.save(book);
	}

	public Optional<Book> updateBook(Long id, Book book) {
		return bookRepository.findById(id)
				.map(existingBook -> {
					book.setId(id);
					return bookRepository.save(book);
				});
	}

	public void deleteBook(Long id) {
		bookRepository.deleteById(id);
	}
}